% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{draw_name: plot the number of number babies with a specified name and gender in France from 1900 to 2015}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{the specified name}

\item{the_sex}{the specified gender}
}
\value{
a line graph plotting the number of babies against year born in a specific year
}
\description{
draw_name: plot the number of number babies with a specified name and gender in France from 1900 to 2015
}
\examples{
\dontrun{
draw_a_name("Diane","F")
}
}
