% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{draw_names: plot the number of new born babies with specified names in France from 1900 to 2015}
\usage{
draw_names(names)
}
\arguments{
\item{names}{a vector of names}
}
\value{
a plot of the number of babies with specified names against year
}
\description{
draw_names: plot the number of new born babies with specified names in France from 1900 to 2015
}
\examples{
\dontrun{
draw_names(c("Diane","Vincent"))
}
}
