% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{multi_excel: reading all sheets of an excel file}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{"the directory of the file"}
}
\value{
a list of all the sheets of the excel file
}
\description{
multi_excel: reading all sheets of an excel file
}
\examples{
\dontrun{
multi_excel("/directory/of/file")
}
}
