% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Importing multiple .csv files}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{directory of the folder}
}
\value{
a list
}
\description{
Importing multiple .csv files
}
\examples{
\dontrun{
my_csv_reader("../../folder/directory")
}
}
