% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_save_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{save_as_csv: a function to save a data frame as a .csv file with built in verification mechanisms for the necessary conditions}
\usage{
save_as_csv(data, output, row.names = FALSE, ...)
}
\arguments{
\item{data}{a data frame}

\item{output}{the directory of the output file, it should end with ".csv"}

\item{row.names}{FALSE by default, this indicates whether the row names of the data is to be written along with the data}

\item{...}{other default settings}
}
\value{
a message saying that the file has been successfully written in the desired location
}
\description{
save_as_csv: a function to save a data frame as a .csv file with built in verification mechanisms for the necessary conditions
}
\examples{
\dontrun{
save_as.csv(iris, "output.csv")
}
}
